new snd folder, no stl
