+++ CMakeLists.txt
@@ -197,6 +197,7 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS 1)
 add_subdirectory (shipdata)
 add_subdirectory (passive)
 add_subdirectory (nutaudet)
+add_subdirectory(SND/MTC)
 add_subdirectory (ecal)
 add_subdirectory (splitcal)
 add_subdirectory (hcal)
