+++ SND/MTC/MTCDetectorLinkDef.h
@@ -1,0 +1,11 @@
+
+#ifdef __CINT__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ class MTCDetector+;
+#pragma link C++ class MtcDetPoint+;
+
+#endif
