+++ geometry/MTC_config.yaml
@@ -7,4 +7,3 @@ MTC:
   nLayers: 45                # Number of sandwich layers
   zPosition: -3343.64        # Global z position (cm)
   fieldY: -1.7               # Magnetic field strength (tesla)
-  isActive: true             # Field is on?
