+++ geometry/geometry_config.py
@@ -70,15 +70,14 @@ if "shieldName" not in globals():
     shieldName = None
 if "SND" not in globals():
     SND = True
-if "SNDDesign" not in globals():
-    SNDDesign = 1
+if "SND_design" not in globals():
+    SND_design = 1
 
 with ConfigRegistry.register_config("basic") as c:
 
     c.DecayVolumeMedium = DecayVolumeMedium
     c.SND = SND
-    c.SNDDesign = SNDDesign
-
+    c.SND_design = SND_design
 
     if not shieldName:
         raise ValueError("shieldName must not be empty!")
