+++ macro/eventDisplay.py
@@ -472,6 +472,7 @@ class DrawTracks(ROOT.FairTask):
                 "strawtubesPoint",
                 "ShipRpcPoint",
                 "TargetPoint",
+                "MtcDetPoint",
                 "TimeDetPoint",
             ]:
                 if not sTree.GetBranch(P):
@@ -1383,6 +1384,10 @@ else:
     mcHits["TargetPoints"] = ROOT.FairMCPointDraw(
         "TargetPoint", ROOT.kRed, ROOT.kFullSquare
     )
+    mcHits["MtcDetPoint"] = ROOT.FairMCPointDraw(
+        "MtcDetPoint", ROOT.kGreen, ROOT.kFullSquare
+    )
+
 
 for x in mcHits:
     fMan.AddTask(mcHits[x])
